requirements.txt
azure-cognitiveservices-language-luis
json
numpy
python-dotenv
scipy
tabulate
transformers